Proyectos Treinta/b2b/tropi_standarization/pyproject.toml
[tool.poetry]
name = "tropi_standarization"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pydantic = "^1.9.1"
google-cloud-secret-manager = "^2.12.0"
google-cloud-bigquery = "^3.3.0"
google-auth = "^2.9.1"
google-oauth = "^1.0.0"
mysql-connector = "^2.2.9"
pandas = "^1.4.3"
db-dtypes = "^1.0.2"
google-api-python-client = "^2.55.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
openpyxl = "^3.0.10"
holidays-co = "^1.0.0"
pandas-gbq = "^0.17.8"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
ipykernel = "^6.15.1"
google-cloud = "^0.34.0"
google-oauth = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
